+++ lib/baby_squeel/table.rb
@@ -37,6 +37,10 @@ module BabySqueel
       end
     end
 
+    def sift(sifter_name, *args)
+      Nodes.wrap @scope.public_send("sift_#{sifter_name}", *args)
+    end
+
     # Alias a table. This is only possible when joining
     # an association explicitly.
     def alias(alias_name)