+++ lib/baby_squeel/active_record.rb
@@ -38,20 +38,10 @@ module BabySqueel
     end
 
     module WhereChain
-      if ::ActiveRecord::VERSION::MAJOR > 4
-        # Constructs Arel for ActiveRecord::Base#where using the DSL.
-        def has(&block)
-          opts = DSL.evaluate(@scope, &block)
-          factory = @scope.send(:where_clause_factory)
-          @scope.where_clause += factory.build(opts, [])
-          @scope
-        end
-      else
-        # Constructs Arel for ActiveRecord::Base#where using the DSL.
-        def has(&block)
-          @scope.where_values += [DSL.evaluate(@scope, &block)]
-          @scope
-        end
+      # Constructs Arel for ActiveRecord::Base#where using the DSL.
+      def has(&block)
+        @scope.where! DSL.evaluate(@scope, &block)
+        @scope
       end
     end
   end