+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## [Unreleased]
+## Unreleased
 
+Nothing yet!
+
+## [0.2.2] - 2015-03-30
 ### Added
 - Support for `group` (`grouping`) and `having` (`when_having`).
 - Support for sifters.
@@ -33,6 +36,7 @@
 ### Added
 - Initial support for selects, orders, wheres, and joins.
 
-[Unreleased]: https://github.com/rzane/baby_squeel/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/rzane/baby_squeel/compare/v0.2.2...HEAD
+[0.2.1]: https://github.com/rzane/baby_squeel/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/rzane/baby_squeel/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/rzane/baby_squeel/compare/v0.1.0...v0.2.0